+++ src/core/client/startup.ts
@@ -52,7 +52,6 @@ import './systems/tasks';
 import './systems/tick';
 import './systems/toolbar';
 import './systems/vehicle';
-import './systems/world';
 
 // Utility
 import './utility/cinematic';
